Refactor Elements to Nodes in Blueprint System
Replaced `Elements` with `Nodes` in the `BlueprintEditor` class, updating the `DependencyProperty` and property type. Modified the `IBlueprintEditor` interface to reflect this change. Updated `BlueprintController` and `BlueprintRenderer` classes to operate on `Node` objects instead of `Element` objects, adjusting methods and rendering logic accordingly. Changed `MainViewModel` to use `IEnumerable<Node>` and updated the `ConnectNodes` method. Updated `MainView.xaml` to bind `Nodes` in `BlueprintEditor` to `ViewModel`. This refactoring aims to better represent the data model and improve functionality.